refactored subsystem custom type parsing
