+++ cmd/knowledge/knowledge.go
@@ -20,13 +20,12 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/cisco-open/fsoc/config"
 	"github.com/cisco-open/fsoc/platform/api"
 )
 
 // Config defines the subsystem configuration under fsoc
 type Config struct {
-	ApiVersion config.ApiVersion `mapstructure:"apiver,omitempty" fsoc-help:"API version to use, e.g., v1"`
+	ApiVersion api.Version `mapstructure:"apiver,omitempty" fsoc-help:"API version to use, e.g., v1"`
 }
 
 var GlobalConfig Config
